microservices/actuate_microservice/requirements.txt
paho-mqtt
kafka-python===2.0.2
paho-mqtt
pytz